Automated Investigation & Blocking Logic for WAF
This is a logic app template solely designed for conducting automated investigations through VirusTotal, CloudMersive and IPQS, and creates a custom fraud/threat score, and decides whether the IP address detected by the algorithm should be blocked or not. It then creates and email and Teams alert which can be fully customised. If the threat score is high or critical, the IP address is automatically added to a custom rule in the WAF using the latest REST API version on Microsoft.